Refactor record control hooks and components to use new record mutations and navigation hooks
