+++ src/features/edit/hooks/useSaveRecord.ts
@@ -2,13 +2,14 @@ import { useSearchParams } from 'react-router-dom';
 
 import { QueryParams } from '@/common/constants/routes.constants';
 import { useModalControls } from '@/common/hooks/useModalControls';
-import { useRecordControls } from '@/common/hooks/useRecordControls';
+
+import { useRecordMutations } from '@/features/resources';
 
 import { useStatusState } from '@/store';
 
 export const useSaveRecord = (primary: boolean) => {
   const { isRecordEdited } = useStatusState(['isRecordEdited']);
-  const { saveRecord } = useRecordControls();
+  const { saveRecord } = useRecordMutations();
   const { isModalOpen, openModal, closeModal } = useModalControls();
   const [searchParams] = useSearchParams();
