+++ src/features/resources/hooks/index.ts
@@ -1,3 +1,5 @@
 export { useRecordGeneration } from './useRecordGeneration';
 export { useProcessedRecordAndSchema } from './useProcessedRecordAndSchema';
 export { useMarcData } from './useMarcData';
+export { useRecordMutations } from './useRecordMutations';
+export { useRecordNavigation } from './useRecordNavigation';
