+++ src/features/resources/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
+export * from './helpers';
 export * from './hooks';
